requirements.txt
blinker==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
Flask==3.0.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.4.0
gunicorn==21.2.0
hub-sdk==0.0.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
Pillow==9.3.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0+cpu
torchvision==0.16.0+cpu
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.4
ultralytics==8.1.5
urllib3==1.26.13
Werkzeug==3.0.1
